# Build time improvements: Eject samples from CRA