--- a/common/config/webpack/sample.webpack.config.js
+++ b/common/config/webpack/sample.webpack.config.js
@@ -8,25 +8,25 @@ const path = require('path');
 const webpack = require('webpack');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
 
-const webpackConfig = (sampleRootDir, env) => ({
+const webpackConfig = (sampleAppDir, env) => ({
   entry: './src/index.tsx',
   ...(env.production || !env.development ? {} : { devtool: 'eval-source-map' }),
   resolve: {
     extensions: ['.ts', '.tsx', '.js'],
     alias: {
       // reference internal packlets src directly for hot reloading when developing
-      '@azure/communication-react': path.resolve(sampleRootDir, '../../packages/communication-react/src'),
-      'react-components': path.resolve(sampleRootDir, '../../packages/react-components/src'),
-      'react-composites': path.resolve(sampleRootDir, '../../packages/react-composites/src'),
-      'chat-stateful-client': path.resolve(sampleRootDir, '../../packages/chat-stateful-client/src'),
-      'chat-component-bindings': path.resolve(sampleRootDir, '../../packages/chat-component-bindings/src'),
-      'calling-stateful-client': path.resolve(sampleRootDir, '../../packages/calling-stateful-client/src'),
-      'calling-component-bindings': path.resolve(sampleRootDir, '../../packages/calling-component-bindings/src'),
-      'acs-ui-common': path.resolve(sampleRootDir, '../../packages/acs-ui-common/src')
+      '@azure/communication-react': path.resolve(sampleAppDir, '../../packages/communication-react/src'),
+      'react-components': path.resolve(sampleAppDir, '../../packages/react-components/src'),
+      'react-composites': path.resolve(sampleAppDir, '../../packages/react-composites/src'),
+      'chat-stateful-client': path.resolve(sampleAppDir, '../../packages/chat-stateful-client/src'),
+      'chat-component-bindings': path.resolve(sampleAppDir, '../../packages/chat-component-bindings/src'),
+      'calling-stateful-client': path.resolve(sampleAppDir, '../../packages/calling-stateful-client/src'),
+      'calling-component-bindings': path.resolve(sampleAppDir, '../../packages/calling-component-bindings/src'),
+      'acs-ui-common': path.resolve(sampleAppDir, '../../packages/acs-ui-common/src')
     }
   },
   output: {
-    path: path.join(sampleRootDir, '/dist'),
+    path: path.join(sampleAppDir, '/dist'),
     filename: 'build.js'
   },
   module: {
@@ -58,8 +58,11 @@ const webpackConfig = (sampleRootDir, env) => ({
     }),
     new webpack.DefinePlugin({
       'process.env.PRODUCTION': env.production || !env.development,
-      'process.env.NAME': JSON.stringify(require(path.resolve(sampleRootDir, './package.json')).name),
-      'process.env.VERSION': JSON.stringify(require(path.resolve(sampleRootDir, './package.json')).version)
+      'process.env.NAME': JSON.stringify(require(path.resolve(sampleAppDir, './package.json')).name),
+      'process.env.VERSION': JSON.stringify(require(path.resolve(sampleAppDir, './package.json')).version)
+    }),
+    new webpack.DefinePlugin({
+      __BUILDTIME__: JSON.stringify(new Date().toLocaleString())
     })
   ],
   devServer: {