--- a/common/config/webpack/sampleapp.webpack.config.js
+++ b/common/config/webpack/sampleapp.webpack.config.js
@@ -41,11 +41,8 @@ const webpackConfig = (sampleAppDir, env) => ({
         use: ['style-loader', 'css-loader']
       },
       {
-        test: /\.(png|jpe?g|gif|svg|eot|ttf|woff|woff2)$/i,
-        loader: 'url-loader',
-        options: {
-          limit: 8192
-        }
+        test: /\.svg/,
+        type: 'asset/inline'
       }
     ]
   },