--- a/packages/api-extractor-fork/bin/api-extractor
+++ b/packages/api-extractor-fork/bin/api-extractor
@@ -1,0 +1,2 @@
+#!/usr/bin/env node
+require('../lib/start.js');