--- a/packages/api-extractor-fork/lib/aedoc/PackageDocComment.d.ts
+++ b/packages/api-extractor-fork/lib/aedoc/PackageDocComment.d.ts
@@ -1,0 +1,9 @@
+import * as ts from 'typescript';
+import { Collector } from '../collector/Collector';
+export declare class PackageDocComment {
+  /**
+   * For the given source file, see if it starts with a TSDoc comment containing the `@packageDocumentation` tag.
+   */
+  static tryFindInSourceFile(sourceFile: ts.SourceFile, collector: Collector): ts.TextRange | undefined;
+}
+//# sourceMappingURL=PackageDocComment.d.ts.map