--- a/packages/api-extractor-fork/lib/start.d.ts
+++ b/packages/api-extractor-fork/lib/start.d.ts
@@ -1,0 +1,2 @@
+export {};
+//# sourceMappingURL=start.d.ts.map