# What
Eject samples from CRA
* Use a barebones webpack (and webpack dev server for development)

Also:
* Upgrade to webpack5 to fix dev server hanging issue
* Fix dependencies that broke build when `rush update --full` performed
* Remove use of preval to get build time and use webpack variable injection instead
* Upgrade rush and pnpm to try fix build issue with api-extractor after rush update full
  * Ended up having to extract api-extractor to a project in out repo - rush+pnpm doesn't handle directly using tarballs, making this PR pretty enormous (but is just the tarball expanded into files)
  * Hopefully we can remove all these very soon:
      ![image](https://user-images.githubusercontent.com/2684369/121909661-0052c900-cce3-11eb-8406-24b2625a1244.png)


# Why
Perf
* We have large CI build times with easiest large gains coming from samples

Dev inner loop improvements
* No longer rely on rush module linking for samples and directly alias to source code
* This massivly improves local build times, more reliable hot reloading, much faster hot reloading and less confusing output in the terminal

# Perf

CI Runs:
| Before | After | Improvement |
| -- | -- | -- |
| 13-16 minutes | 8-10minutes | -5-6 minutes |
| ![image](https://user-images.githubusercontent.com/2684369/121597939-1c5e1e00-c9f6-11eb-98fb-dd052fd3e7f7.png) | ![image](https://user-images.githubusercontent.com/2684369/121597827-00f31300-c9f6-11eb-9372-e2b8d6ccce7b.png) |


Local Tests for Calling Sample:
| | Before | After | Improvement |
| -- | -- | -- | -- |
| Production Build (dependencies already built) | 63s | 51s | -12s |
| Production Build (dependencies not built) | 5min 4s | 51s | -4min 15s |
| `npm run start` (dependencies already built) (until homepage loaded) | 2min 17s | 27s | -1min 50s |
| `npm run start` (dependencies not built) (until homepage loaded) | 3min 56s | 27s | -2min 29s |
| Hot reload (change made in `react-components` dir) | 20 seconds (+ other packlets still recompiling) | 4s | -16s |

Profiler still crashes 😥

# How Tested
* Samples working locally
* Deployed calling sample to hosted azure web app - tested desktop edge and ios safari