@JamesBurnside 2021-06-11
> The build time improvements are amazing. Bravo!
> 
> I'm a bit conflicted about sample apps using the uncompiled package sources though.
> Is the source mapping also required for build time improvements, or is it only for hot reloading?
> 
> The concern would be some build related issues slipping through the cracks / the samples not quite working if they code is copied as-is to an independent project (which customers are expected to do to kickstart their apps).

Just for hot reloading (though build time will also be affected as first the dependencies will need built then the sample app webpack run).

I agree with the concerns but I think the solution is to move samples to using the public npm package (either latest of alpha, we can decide), then for local dev you run a script that points the sample to the local communication-react folder to support hot reloading.