@JamesBurnside 2021-06-11
> Land this already so I can speed up my edit-compile-test cycle for the day :D

I don't feel hesitant anymore having to run the samples! It's really nice 😅