@anjulgarg 2021-06-11
Thanks for landing this @JamesBurnside 
Did you have to remove polyfills and shims in webpack config?
I remember the original ejected webpack being a very large file. 
Are we removing something that might cause our sample to stop working as expected on a specific browser?
If yes, I think we should test the generated sample thoroughly on each major browser before we merge this in.